updated for multiple files, RXEN pin
