+++ README.md
@@ -3,3 +3,8 @@
 Library for communicating with Modbus devices over RS232/USB/485 via RTU protocol
 
 Ported to Particle by Paul Kourany, 2015
+
+# Update May 2016
+
+- Added in a RXEN pin
+- Changed how RXEN and TXEN works
